refactor: rename internal/snyk to internal/sarif
The package parses generic SARIF v2.1.0 (used by Snyk, Checkov, Trivy,
Semgrep, etc.), not Snyk-specifically. Rename to reflect the actual
scope ahead of adding a first-class `kosli attest sarif` command that
will share this parser.

No behaviour change: the snyk attestation wire format keeps the
`snyk_results` JSON key, and the snyk command continues to invoke the
shared parser via its new name (`sarif.ProcessSarifResultFile`).
